Enable runtime changes to tcmalloc sampling parameter
Before this diff, the tcmalloc sampling interval was defined by
FLAGS_tcmalloc_sample_parameter, which had no atomic guards around
modifying it. (NB: This uses commandlineflags.h defined in this
repository, not gflags). This commit changes it to a static variable and
adds an API to MallocExtension so that it can be safely get/set at
runtime.